\name{distobj-package}
\alias{distobj-package}
\alias{distobj}
\docType{package}
\title{
\packageTitle{distobj}
}
\description{
\packageDescription{distobj}
}
\details{
\packageDESCRIPTION{distobj}
\packageIndices{distobj}

A comprehensive system for the creation and control of distributed objects for
standard R usage in statistical modelling.
}
\author{
\packageAuthor{distobj}

Maintainer: \packageMaintainer{distobj}
}
\keyword{ package }
