\name{distObj-package}
\alias{distObj-package}
\alias{distObj}
\docType{package}
\title{
\packageTitle{distObj}
}
\description{
\packageDescription{distObj}
}
\details{
\packageDESCRIPTION{distObj}
\packageIndices{distObj}

A comprehensive system for the creation and control of distributed objects for
standard R usage in statistical modelling.
}
\author{
\packageAuthor{distObj}

Maintainer: \packageMaintainer{distObj}
}
\keyword{ package }
