\name{largeScaleR-package}
\alias{largeScaleR-package}
\alias{largeScaleR}
\docType{package}
\title{
\packageTitle{largeScaleR}
}
\description{
\packageDescription{largeScaleR}
}
\details{
\packageDESCRIPTION{largeScaleR}
\packageIndices{largeScaleR}

A comprehensive system for the creation and control of distributed objects for
standard R usage in statistical modelling.
}
\author{
\packageAuthor{largeScaleR}

Maintainer: \packageMaintainer{largeScaleR}
}
\keyword{ package }
